requirements.txt
trakt.py
pycliarr==1.0.14
schedule